+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Release 1.2.0
-- Added support for meshes.
+- Added support for costume meshes.
+- Added support for custom promos.
 - Better character exporting and importing. Note: This will break compatibility with older versions of the mod. Future versions will aim to be backwards compatible.
 - Exported characters now have the `.character` extension instead of `.json`, and `Meta` and `ContentMappings` now have the `.meta` and `.mappings` extensions respectively.